streaming/vision/ -> examples/vision/.
