+++ examples/text/__init__.py
@@ -1,4 +1,4 @@
 # Copyright 2023 MosaicML Streaming authors
 # SPDX-License-Identifier: Apache-2.0
 
-"""Example text streaming dataset."""
+"""Example text streaming datasets."""
