+++ examples/vision/__init__.py
@@ -1,4 +1,4 @@
 # Copyright 2023 MosaicML Streaming authors
 # SPDX-License-Identifier: Apache-2.0
 
-"""Data conversion scripts for Computer Vision."""
+"""Example computer vision streaming datasets."""
