+++ examples/vision/ade20k/__init__.py
@@ -1,0 +1,1 @@
+"""ADE20k streaming dataset example."""
