+++ examples/vision/ade20k/read.py
