+++ examples/vision/ade20k/write.py
