+++ examples/vision/cifar10/__I
@@ -1,0 +1,1 @@
+"""CIFAR10 streaming dataset example."""
