+++ examples/vision/cifar10/read.py
