+++ examples/vision/cifar10/write.py
