+++ examples/vision/cifar10/write_fake.py
