+++ examples/vision/coco/__init__.py
@@ -1,0 +1,1 @@
+"""COCO streaming dataset example."""
