+++ examples/vision/coco/read.py
