+++ examples/vision/coco/write.py
