+++ examples/vision/imagenet/__init__.py
@@ -1,0 +1,1 @@
+"""ImageNet streaming dataset example."""
