+++ examples/vision/imagenet/write.py
